% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acousticBatch.R
\name{acousticBatch}
\alias{acousticBatch}
\title{Extract acoustic features in batch for multiple files in a directory}
\usage{
acousticBatch(inputdir, outputdir, ...)
}
\description{
Extract acoustic features in batch for multiple files in a directory
}
