% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audioSummary2020.R
\name{audioSummary2020}
\alias{audioSummary2020}
\title{Get a series of summary statistics for a single file}
\usage{
audioSummary2020(wave, threshold, window_length, overlap, plot = F)
}
\arguments{
\item{wave}{Wave object. It may be imported using `tuneR`}

\item{threshold}{Amplitude threshold for separating signal from noise, in %.}

\item{window_length}{Length, in number of samples, of each spectral window.}

\item{plot}{Should acoustic features be plotted?}

\item{overalp}{Percentage of overlap between spectral windows.}
}
\description{
Get a series of summary statistics for a single file
}
