% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv.R
\name{dv}
\alias{dv}
\title{Ratio of amplitudes of pulses and interpulse intervals}
\usage{
dv(sound, ...)
}
\arguments{
\item{sound}{Wave object.}

\item{...}{Further arguments passed to `seewave::timer()`.}
}
\value{
A vector with the ratio of amplitude of each interpulse interval
and its preceding pulse.
}
\description{
Ratio of amplitudes of pulses and interpulse intervals
}
\details{
This script calculates the ratio of amplitudes of pulses and
interpulse intervals according to Ryan & Sullivan (1989). The root mean
squares for each pulse and interpulse interval are calculated. Then the
rms of each interpulse interval is divided by the rms of the preceding
pulse, and the result is substracted from 1. The last pulse is not
considered in the calculation because by definition it does not precede
any other interpulse interval.
}
