% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{intervals}
\alias{intervals}
\title{Durations of syllables and intersyllable intervals}
\usage{
intervals(sound, labels = NULL, ...)
}
\arguments{
\item{sound}{Wave object.}

\item{labels}{File containing the timestamps of the syllables.}

\item{...}{Further arguments passed to `seewave::timer()`.}
}
\description{
Durations of syllables and intersyllable intervals
}
