% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acousticBatch.R
\name{acousticBatch}
\alias{acousticBatch}
\title{Extract acoustic features in batch for multiple files in a directory}
\usage{
acousticBatch(inputdir, outputdir, ...)
}
\arguments{
\item{inputdir}{Folder containing wave files.}

\item{outputdir}{Folder where to write results.}

\item{...}{Arguments to be passed to `audioSummary2020`.}
}
\description{
Extract acoustic features in batch for multiple files in a directory
}
